clean up classes in the same file
