+++ src/backend/TrafficCourts/Citizen.Service/Services/Tickets/Search/TicketSearchErrorException.cs
@@ -1,10 +1,10 @@
-﻿namespace TrafficCourts.Citizen.Service.Services.Tickets.Search;
+﻿using System.Runtime.Serialization;
+
+namespace TrafficCourts.Citizen.Service.Services.Tickets.Search;
 
 [Serializable]
 public class TicketSearchErrorException : Exception
 {
     public TicketSearchErrorException(string message, Exception inner) : base(message, inner) { }
-    protected TicketSearchErrorException(
-      System.Runtime.Serialization.SerializationInfo info,
-      System.Runtime.Serialization.StreamingContext context) : base(info, context) { }
+    protected TicketSearchErrorException(SerializationInfo info, StreamingContext context) : base(info, context) { }
 }
